separate EthSimBundle namespace
